Registro ok sin considerar imagen
