+++ views/partials/header.ejs
@@ -97,7 +97,7 @@
                             <button type="submit" class="btn btn-light mt-3">Iniciar Sesión
                             <i class="fa-solid fa-paw"></i></button>
                         </form>
-                        <p class="pt-3 text-center">¿Aún no te has registrado? <a class="text-decoration-none" href="#register" 
+                        <p class="pt-3 text-center">¿Aún no te has registrado? <a class="text-decoration-none" href="/usuario/registro" 
                         data-bs-toggle="modal" data-bs-dismiss="modal">¡Hazlo ya!</a></p>
                     </div>
                 </div>
